requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.7
astroid==2.4.2
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.6
django-cleanup==4.0.0
django-crispy-forms==1.8.0
django-filter==2.3.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient @ file:///D:/collegeDB%20Project/mysqlclient-1.4.6-cp38-cp38-win32.whl
oauthlib==3.1.0
passlib==1.7.2
Pillow==6.2.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==2.0.0
toml==0.10.1
urllib3==1.25.7
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
